+++ common/src/main/scala/it/codingjam/lagioconda/FitnessCalculator.scala
@@ -1,0 +1,5 @@
+package it.codingjam.lagioconda
+
+case class FitnessCalculator(fitnessService: ActorSelection) {
+
+}
