+++ common/src/main/scala/it/codingjam/lagioconda/population/Individual.scala
@@ -1,4 +1,4 @@
-package it.codingjam.lagioconda.models
+package it.codingjam.lagioconda.population
 
 import it.codingjam.lagioconda.ga.Chromosome
