+++ common/src/main/scala/it/codingjam/lagioconda/population/package.scala
@@ -1,0 +1,5 @@
+package it.codingjam.lagioconda
+
+package object population {
+
+}
