+++ frontend/app/actors/PopulationActor.scala
@@ -10,7 +10,7 @@ import it.codingjam.lagioconda.actors.SocketActor.{GenerationRan, PopulationGene
 import it.codingjam.lagioconda.conversions.ChromosomeToBufferedImage
 import it.codingjam.lagioconda.fitness.ByteComparisonFitness
 import it.codingjam.lagioconda.ga.{MutationPointLike, _}
-import it.codingjam.lagioconda.models.Individual
+import it.codingjam.lagioconda.population.Individual
 import it.codingjam.lagioconda.protocol.Messages.IndividualInfo
 import it.codingjam.lagioconda.{Configuration, ImageDimensions, Population, WheelSelection}
 import org.apache.commons.codec.binary.Base64OutputStream
