+++ frontend/app/models/EliteSelection.scala
@@ -1,7 +1,7 @@
 package models
 
+import it.codingjam.lagioconda.population.Individual
 import it.codingjam.lagioconda.{Population, SelectionFunction}
-import it.codingjam.lagioconda.models.Individual
 
 class EliteSelection extends SelectionFunction {
   override def select(population: Population): Individual = population.randomElite
