+++ frontend/app/models/SelectionFunction.scala
@@ -1,6 +1,6 @@
 package it.codingjam.lagioconda
 
-import it.codingjam.lagioconda.models.Individual
+import it.codingjam.lagioconda.population.Individual
 
 trait SelectionFunction {
   def select(population: Population): Individual
